+++ tests/test_catalog.py
@@ -1238,14 +1238,24 @@ class TestCatalog:
             )
         )
 
-        html_link = catalog.get_single_link("search")
+        html_link = catalog.get_single_link(rel="search")
         assert html_link is not None
         assert html_link.href == "./search.html"
-        json_link = catalog.get_single_link("search", media_type="application/geo+json")
+        html_link = catalog.get_single_link(media_type="text/html")
+        assert html_link is not None
+        assert html_link.href == "./search.html"
+        json_link = catalog.get_single_link(
+            rel="search", media_type="application/geo+json"
+        )
         assert json_link is not None
         assert json_link.href == "./search.json"
+        no_link = catalog.get_single_link(rel="via")
+        assert no_link is None
+        first_link = catalog.get_single_link()
+        assert first_link is not None
+        assert first_link.rel == "self"
 
-    def test_get_links_media_type(self) -> None:
+    def test_get_links(self) -> None:
         catalog = TestCases.case_1()
 
         catalog.links.append(
@@ -1256,7 +1266,13 @@ class TestCatalog:
                 rel="search", target="./search.json", media_type="application/geo+json"
             )
         )
-        assert len(catalog.get_links("search", media_type="application/geo+json")) == 1
+        assert (
+            len(catalog.get_links(rel="search", media_type="application/geo+json")) == 1
+        )
+        assert len(catalog.get_links(media_type="text/html")) == 1
+        assert len(catalog.get_links(rel="search")) == 2
+        assert len(catalog.get_links(rel="via")) == 0
+        assert len(catalog.get_links()) == 6
 
     def test_to_dict_no_self_href(self) -> None:
         catalog = Catalog(id="an-id", description="A test Catalog")